pyproject.toml
[tool.poetry]
name = "skypro-course-beta"
version = "0.1.0"
description = ""
authors = ["Denis Antropov <de94onisus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
requests = "^2.31.0"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pandas-stubs = "^2.1.1.230928"
types-requests = "^2.31.0.10"


[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # исключить несколько общих каталогов в
    | \.git          # корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # также отдельно исключить файл с именем foo.py в
                     # корне проекта
)
'''

[tool.isort]
line_length = 119